backend/requirements.txt
###### uvicorn app.main:app --reload #####
###### alembic revision --autogenerate -m "command" #####
###### alembic upgrade head #####
###### (ถ้ามี venv ให้ activate ก่อน) ######
###### venv\Scripts\activate ######
###### alembic current ######
###### pip install -r requirements.txt ######
###### [alembic]
# script_location = alembic
###### cd backend
# alembic -c alembic.ini current
# alembic -c alembic.ini upgrade head



annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.122.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
